Refactored isIP and isMAC to seperate file
