+++ app/imports/server/api/endpoints.test.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { chai } from 'meteor/practicalmeteor:chai';
-import { getEntries, isMAC, isIP } from '../../../server/api/endpoints.js';
+import { getEntries, clearExpiredEntries } from '../../../server/api/endpoints.js';
 
 var assert = chai.assert,
     expect = chai.expect;
@@ -27,78 +27,6 @@ const processedThreeEntries = [{
     IP: '192.168.1.18',
     MAC: '3C:15:C2:CC:DC:2A'
 }];
-
-describe('MAC regex', function() {
-    it('accepts dash notation MAC', function() {
-        assert(isMAC('00-26-BB-07-54-80'));
-    });
-
-    it('accepts column notation MAC', function() {
-        assert(isMAC('8C:3A:E3:93:83:93'));
-    });
-
-    it('rejects too short MAC', function() {
-        assert(!isMAC('00-26-BB-07-54'));
-    });
-
-    it('rejects too long MAC', function() {
-        assert(!isMAC('00:26:BB:07:54:34:FD'));
-    });
-
-    it('rejects lower case MAC', function() {
-        assert(!isMAC('00-26-bb-07-54-80'));
-    });
-
-    it('rejects empty MAC', function() {
-        assert(!isMAC(''));
-    });
-
-    it('rejects MAC without separation', function() {
-        assert(!isMAC('0026BB075480'));
-    });
-
-    it('only accepts A-F as characters', function() {
-        assert(!isMAC('00-26-RB-07-54-80'));
-    });
-});
-
-describe('IP regex', function() {
-    it('accepts correct IP', function() {
-        assert(isIP('192.168.1.50'));
-    });
-
-    it('rejects too long IP', function() {
-        assert(!isIP('192.168.1.50.23'));
-    });
-
-    it('rejects too short IP', function() {
-        assert(!isIP('192.168.1'));
-    });
-
-    it('rejects too high value IP', function() {
-        assert(!isIP('192.521.1.50'));
-    });
-
-    it('rejects IP with characters', function() {
-        assert(!isIP('192.168.1a.50'));
-    });
-
-    it('rejects IP with columns', function() {
-        assert(!isIP('192:168:1:50'));
-    });
-
-    it('rejects IP with dashes', function() {
-        assert(!isIP('192-168-1-50'));
-    });
-
-    it('rejects IP with no separation', function() {
-        assert(!isIP('192168150'));
-    });
-
-    it('rejects empty IP', function() {
-        assert(!isIP(''));
-    });
-});
 
 describe('arpTable', function() {
     it('processes correct arpTables', function() {
@@ -145,4 +73,10 @@ describe('arpTable', function() {
         getEntries(ReUsedIPandMAC);
     });
 
+});
+
+describe('clearExpiredEntries', function() {
+    it('removes entries that have expired', function() {
+
+    });
 });
