+++ app/server/api/endpoints.js
@@ -1,3 +1,5 @@
+import { isMAC, isIP } from '../regexes.js';
+
 RouterApi = new Restivus({
     apiPath: 'routerapi/',
     version: 'v1'
@@ -47,14 +49,6 @@ var getEntries = function(arpTable) {
 
     return entries;
 };
-
-var isIP = function(field) {
-    return new RegExp("^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$").test(field);
-}
-
-var isMAC = function(field) {
-    return new RegExp("^([0-9A-F]{2}[:-]){5}([0-9A-F]{2})$").test(field);
-}
 
 var clearExpiredEntries = function(entries) {
     // The user is not connected to the router anymore
